+++ libcaf_core/caf/json_value.hpp
@@ -19,6 +19,7 @@ class CAF_CORE_EXPORT json_value {
 public:
   // -- constructors, destructors, and assignment operators --------------------
 
+  /// Creates a @c null JSON value.
   json_value() noexcept : val_(detail::json::null_value()) {
     // nop
   }
@@ -39,6 +40,8 @@ public:
 
   // -- factories --------------------------------------------------------------
 
+  /// Creates an undefined JSON value. This special state usually indicates that
+  /// a previous key lookup failed.
   static json_value undefined() noexcept {
     return json_value{detail::json::undefined_value(), nullptr};
   }
@@ -57,19 +60,19 @@ public:
   }
 
   /// Checks whether the value is an @c int64_t.
-  bool is_integer() const noexcept {
-    return val_->is_integer();
-  }
+  bool is_integer() const noexcept;
+
+  /// Checks whether the value is an @c uint64_t.
+  bool is_unsigned() const noexcept;
 
   /// Checks whether the value is a @c double.
   bool is_double() const noexcept {
     return val_->is_double();
   }
 
-  /// Checks whether the value is a number, i.e., an @c int64_t or a @c double.
-  bool is_number() const noexcept {
-    return is_integer() || is_double();
-  }
+  /// Checks whether the value is a number, i.e., an @c int64_t, a @c uint64_t
+  /// or a @c double.
+  bool is_number() const noexcept;
 
   /// Checks whether the value is a @c bool.
   bool is_bool() const noexcept {
@@ -93,26 +96,49 @@ public:
 
   // -- conversion -------------------------------------------------------------
 
+  /// Converts the value to an @c int64_t or returns @p fallback if the value
+  /// is not a valid number.
   int64_t to_integer(int64_t fallback = 0) const;
 
+  /// Converts the value to an @c uint64_t or returns @p fallback if the value
+  /// is not a valid number.
+  uint64_t to_unsigned(uint64_t fallback = 0) const;
+
+  /// Converts the value to a @c double or returns @p fallback if the value is
+  /// not a valid number.
   double to_double(double fallback = 0.0) const;
 
+  /// Converts the value to a @c bool or returns @p fallback if the value is
+  /// not a valid boolean.
   bool to_bool(bool fallback = false) const;
 
+  /// Returns the value as a JSON string or an empty string if the value is not
+  /// a string.
   std::string_view to_string() const;
 
+  /// Returns the value as a JSON string or @p fallback if the value is not a
+  /// string.
   std::string_view to_string(std::string_view fallback) const;
 
+  /// Returns the value as a JSON array or an empty array if the value is not
+  /// an array.
   json_array to_array() const;
 
+  /// Returns the value as a JSON array or @p fallback if the value is not an
+  /// array.
   json_array to_array(json_array fallback) const;
 
+  /// Returns the value as a JSON object or an empty object if the value is not
+  /// an object.
   json_object to_object() const;
 
+  /// Returns the value as a JSON object or @p fallback if the value is not an
+  /// object.
   json_object to_object(json_object fallback) const;
 
   // -- comparison -------------------------------------------------------------
 
+  /// Compares two JSON values for equality.
   bool equal_to(const json_value& other) const noexcept;
 
   // -- parsing ----------------------------------------------------------------
@@ -143,6 +169,7 @@ public:
 
   // -- printing ---------------------------------------------------------------
 
+  /// Prints the JSON value to @p buf.
   template <class Buffer>
   void print_to(Buffer& buf, size_t indentation_factor = 0) const {
     detail::json::print_to(buf, *val_, indentation_factor);
@@ -150,6 +177,7 @@ public:
 
   // -- serialization ----------------------------------------------------------
 
+  /// Applies the `Inspector` to the JSON value `val`.
   template <class Inspector>
   friend bool inspect(Inspector& inspector, json_value& val) {
     if constexpr (Inspector::is_loading) {